move onCreate(db) to ODistributed plugin
